tidy up javascript
