+++ app/javascript/components/aos.js
@@ -1,0 +1,8 @@
+import AOS from 'aos';
+import 'aos/dist/aos.css';
+
+const initAos = () => {
+  AOS.init();
+}
+
+export { initAos }
