+++ app/javascript/components/filter_listener.js
@@ -1,43 +1,45 @@
 import Rails from "@rails/ujs"
 
-  const greatListener = () => {
+const greatListener = () => {
+  const expertisesForm = document.getElementById("Expertises");
+  const technologiesForm = document.getElementById("Technologies");
+  const seniorityForm = document.getElementById("Seniority");
+  const form = document.getElementById("filter-form");
 
-  const ExpertisesForm = document.getElementById("Expertises");
-  ExpertisesForm.addEventListener('change', (event) => {
-    // event.preventDefault();
-    submitAjaxForm();
-  });
+  if (expertisesForm) {
 
-  const TechnologiesForm = document.getElementById("Technologies");
-  TechnologiesForm.addEventListener('change', (event) => {
-    // event.preventDefault();
-    submitAjaxForm();
-  });
+    expertisesForm.addEventListener('change', (event) => {
+      // event.preventDefault();
+      submitAjaxForm();
+    });
 
-  const SeniorityForm = document.getElementById("Seniority");
-  SeniorityForm.addEventListener('change', (event) => {
-    // event.preventDefault();
-    submitAjaxForm();
-  });
+    technologiesForm.addEventListener('change', (event) => {
+      // event.preventDefault();
+      submitAjaxForm();
+    });
 
-  const submitAjaxForm = () => {
-    const form = document.getElementById("filter-form");
+    seniorityForm.addEventListener('change', (event) => {
+      // event.preventDefault();
+      submitAjaxForm();
+    });
 
-    let formData = new FormData(form)
-    // Rails ajax corresponding to fetch
-    Rails.ajax({
-      url: form.action,
-      type: "post",
-      data: formData,
-      success: function(data) {
-        console.log(data)
-      },
-      error: function(data) {
-        console.log(data)
-      }
-    })
+    const submitAjaxForm = () => {
+      let formData = new FormData(form)
+      // Rails ajax corresponding to fetch
+      Rails.ajax({
+        url: form.action,
+        type: "post",
+        data: formData,
+        success: function(data) {
+          console.log(data)
+        },
+        error: function(data) {
+          console.log(data)
+        }
+      })
+    }
+
   }
-
 }
 
-export {greatListener}
+export { greatListener }
