+++ app/javascript/components/pure_counter.js
@@ -1,0 +1,7 @@
+import PureCounter from '@srexi/purecounterjs/js/purecounter'
+
+const initPureCounter = () => {
+  const pure = new PureCounter
+}
+
+export { initPureCounter }
