+++ app/javascript/components/validate.js
@@ -1,0 +1,5 @@
+const initValidate = () => {
+  var validate = require("validate.js");
+}
+
+export { initValidate }
